+++ tests/unit/plugins/module_utils/mock_nfs_global_settings_api.py
@@ -1,0 +1,43 @@
+# Copyright: (c) 2023, Dell Technologies
+
+# Apache License version 2.0 (see MODULE-LICENSE or http://www.apache.org/licenses/LICENSE-2.0.txt)
+
+"""Mock Api response for Unit tests of NFS global settings module on PowerScale"""
+
+from __future__ import (absolute_import, division, print_function)
+
+__metaclass__ = type
+
+
+class MockNFSGlobalSettingsApi:
+    MODULE_UTILS_PATH = 'ansible_collections.dellemc.powerscale.plugins.module_utils.storage.dell.utils'
+
+    NFS_GLOBAL_COMMON_ARGS = {
+        "onefs_host": "**.***.**.***",
+        "nfsv3_enabled": None,
+        "nfsv3": None,
+        "nfsv4": None,
+        "rpc_maxthreads": 20,
+        "rpc_minthreads": 17,
+        "rquota_enabled": None,
+        "service": None
+    }
+    GET_NFS_GLOBAL_RESPONSE = {
+        "nfsv3_enabled": False,
+        "nfsv3_rdma_enabled": True,
+        "nfsv40_enabled": True,
+        "nfsv41_enabled": True,
+        "nfsv42_enabled": False,
+        "nfsv4_enabled": True,
+        "rpc_maxthreads": 20,
+        "rpc_minthreads": 17,
+        "rquota_enabled": True,
+        "service": True
+    }
+
+    @staticmethod
+    def get_nfs_global_settings_exception_response(response_type):
+        if response_type == 'get_details_exception':
+            return "Got error SDK Error message while getting NFS global setings details "
+        elif response_type == 'update_exception':
+            return "Modify NFS global settings failed with error: SDK Error message"