+++ CHANGELOG.rst
@@ -4,15 +4,32 @@ Dellemc.Powerscale Change Logs
 
 .. contents:: Topics
 
+
+v2.2.0
+======
+
+Minor Changes
+-------------
+
+- Added support for listing NFS default settings, NFS global settings and NFS zone settings in Info module.
+- Added support for specifying the users and groups to which non-root and root clients are mapped in nfs module.
+
+New Modules
+-----------
+
+- dellemc.powerscale.nfs_default_settings - Get details and modify NFS default settings.
+- dellemc.powerscale.nfs_global_settings - Get details and modify NFS global settings.
+- dellemc.powerscale.nfs_zone_settings - Get details and modify NFS zone settings.
+
 v2.1.0
 ======
 
 Minor Changes
 -------------
 
+- Added support for SmartConnect zone alaises(DNS names) in network pool module.
 - Added support for deleting an access zone and reordering the authentication providers in access zone module.
 - Added support for service principal names(SPN) in AD module.
-- Added support for SmartConnect zone alaises(DNS names) in network pool module.
 
 New Modules
 -----------
@@ -25,10 +42,10 @@ v2.0.0
 Minor Changes
 -------------
 
-- Added support for ignoring unresolvable hosts for NFS Export.
-- Added support for everyone user in filesystem module.
-- Added support for listing LDAP auth providers and user mapping rules in Info module.
 - Added support for PowerScale OneFS 9.5 Islander release.
+- Added support for everyone user in filesystem module.
+- Added support for ignoring unresolvable hosts for NFS Export.
+- Added support for listing LDAP auth providers and user mapping rules in Info module.
 - For the execution of the PowerScale Ansible modules, python library "isilon-sdk" needs to be installed.
 
 New Modules